refactor: extract basic component
