+++ packages/components/src/attachments/components/index.ts
@@ -1,4 +1,2 @@
-export { default as DropZoneOverlay } from './DropZoneOverlay.vue'
-export { default as AttachmentList } from './AttachmentList.vue'
 export { default as FileCard } from './FileCard.vue'
 export { default as ImagePreview } from './ImagePreview.vue'
