+++ packages/components/src/attachments/composables/index.ts
@@ -1,7 +1,6 @@
-import { useDragDrop } from './useDragDrop'
 import { useFileType } from './useFileType'
-import { useImagePreview } from './useImagePreview'
-import { useUpload } from './useUpload'
 import { useFileCard } from './useFileCard'
+import { useImagePreview } from './useImagePreview'
+import { useListType } from './useListType'
 
-export { useDragDrop, useFileType, useImagePreview, useUpload, useFileCard }
+export { useFileType, useFileCard, useImagePreview, useListType }
