+++ packages/components/src/attachments/index.vue
@@ -1,14 +1,13 @@
 <script setup lang="ts">
-import { watch, ref, computed } from 'vue'
-import { useFileDialog } from '@vueuse/core'
-import { AttachmentList, DropZoneOverlay } from './components'
-import { useDragDrop, useFileType, useImagePreview, useUpload } from './composables'
-import { AttachmentsEmits, AttachmentsProps, Attachment, PlaceholderConfig } from './index.type'
+import { watch, ref } from 'vue'
+import { useImagePreview, useListType } from './composables'
+import { AttachmentsEmits, AttachmentsProps, Attachment } from './index.type'
+import FileCard from './components/FileCard.vue'
 import './index.less'
 
 const props = withDefaults(defineProps<AttachmentsProps>(), {
-  autoUpload: true,
-  multiple: true,
+  overflow: 'wrap',
+  listType: 'auto', // 默认自动检测
 })
 
 const emit = defineEmits<AttachmentsEmits>()
@@ -16,59 +15,11 @@ const emit = defineEmits<AttachmentsEmits>()
 // 文件列表管理
 const fileList = ref<Attachment[]>(props.items || [])
 
-// 使用文件类型工具
-const { createAttachments } = useFileType()
-
-// 使用上传工具
-const { uploadFile, retryUpload } = useUpload(fileList, {
-  action: props.action,
-  headers: props.headers,
-  data: props.data,
-  withCredentials: props.withCredentials,
-  beforeUpload: props.beforeUpload,
-  customRequest: props.customRequest,
-})
-
-// 处理文件上传
-const handleFilesDropped = (droppedFiles: File[]) => {
-  if (props.disabled) return
-
-  const newFiles = createAttachments(droppedFiles)
-  fileList.value = [...fileList.value, ...newFiles]
-
-  // 触发事件
-  emit('files-dropped', newFiles)
-  emit('update:items', fileList.value)
-
-  // 如果设置了自动上传，则上传新文件
-  if (props.autoUpload !== false) {
-    newFiles.forEach((file) => uploadFile(file))
-  }
-}
-
-// 图片拖拽逻辑
-const { dropZoneRef, dragState, isDragFullscreen } = useDragDrop(
-  {
-    onDrop: handleFilesDropped,
-  },
-  props,
-)
+// 自动检测 listType
+const { actualListType } = useListType(fileList, props.listType)
 
 // 图片预览逻辑
 const { handlePreview, renderPreview } = useImagePreview(fileList, emit, { enableDownload: true })
-
-const { open: handleTriggerSelect, onChange } = useFileDialog({
-  multiple: props.multiple,
-  accept: props.accept,
-})
-
-// 监听文件选择变化
-onChange((selectedFiles) => {
-  if (selectedFiles && selectedFiles.length > 0) {
-    const filesArray = Array.from(selectedFiles)
-    handleFilesDropped(filesArray)
-  }
-})
 
 // 移除文件
 function handleRemove(file: Attachment) {
@@ -94,7 +45,6 @@ function handleDownload(file: Attachment) {
 
 // 重试上传
 function handleRetry(file: Attachment) {
-  retryUpload(file)
   emit('file-retry', file)
 }
 
@@ -114,73 +64,34 @@ watch(
   },
   { deep: true },
 )
-
-const clearAllAttachments = () => {
-  fileList.value.forEach((file) => {
-    if (file.previewUrl && file.previewUrl.startsWith('blob:')) {
-      URL.revokeObjectURL(file.previewUrl)
-    }
-  })
-
-  fileList.value = []
-  emit('update:items', fileList.value)
-}
-
-// 暴露方法给外部调用
-defineExpose({
-  clearFiles: clearAllAttachments,
-  getFiles: () => fileList.value,
-  getFileCount: () => fileList.value.length,
-  hasFiles: () => fileList.value.length > 0,
-  addFiles: handleFilesDropped,
-})
-
-const dropZoneConfig = computed<PlaceholderConfig | undefined>(() => {
-  if (typeof props.placeholder === 'function') {
-    return props.placeholder('drop')
-  }
-  return undefined
-})
 </script>
 
 <template>
-  <div class="tr-attachments rootClass" :style="styles?.root">
-    <div ref="dropZoneRef" class="tr-attachments__dropzone" @click="handleTriggerSelect()">
-      <AttachmentList
-        v-if="fileList.length > 0"
-        :files="fileList"
-        :list-type="listType"
+  <div class="tr-attachments">
+    <div
+      v-if="fileList.length > 0"
+      class="tr-attachments__file-list"
+      :class="`tr-attachments__file-list--${overflow}`"
+      @click.stop
+    >
+      <FileCard
+        v-for="file in fileList"
+        :key="file.uid"
+        :file="file"
+        :list-type="actualListType"
         :file-icons="fileIcons"
         :disabled="disabled"
-        :styles="styles"
         :status-type="statusType"
         :custom-actions="customActions"
-        :overflow="overflow"
+        :show-status="true"
         @remove="handleRemove"
         @preview="handlePreview"
         @download="handleDownload"
         @retry="handleRetry"
+        @file-preview="handlePreview"
         @action="handleAction"
       />
-
-      <!-- 空状态 -->
-      <div v-else class="tr-attachments__empty">
-        <img class="tr-attachments__empty-icon" src="./../assets/svgs/add-file.svg" />
-        <div class="tr-attachments__empty-text">暂无文件</div>
-        <div class="tr-attachments__empty-hint">将文件拖拽到此处，或点击上传</div>
-      </div>
     </div>
-
-    <!-- 全屏拖拽遮罩 -->
-    <DropZoneOverlay
-      :visible="dragState.active"
-      :fullscreen="isDragFullscreen"
-      :icon="dropZoneConfig?.icon"
-      :title="dropZoneConfig?.title"
-      :description="dropZoneConfig?.description"
-      :config="typeof drag === 'object' ? drag.overlay : undefined"
-      :style="styles?.overlay"
-    />
 
     <!-- 图片预览组件 -->
     <Component :is="renderPreview()" />
