+++ packages/components/src/attachments/vars.less
@@ -13,9 +13,9 @@
 
     // 布局和尺寸变量
     --tr-attachments-border-radius: 8px;
-    --tr-attachments-card-height: 58px;
+    --tr-attachments-card-height: 56px;
     --tr-attachments-card-width: 192px;
-    --tr-attachments-icon-size: 42px;
+    --tr-attachments-icon-size: 40px;
     --tr-attachments-close-button-size: 14px;
 
     // 间距变量
